refactor: move credit hook to user feature as it is used outside of the home feature and is about the user
