+++ src/features/bookOffer/pages/BookingConfirmation.tsx
@@ -2,9 +2,9 @@ import { useNavigation, useRoute } from '@react-navigation/native'
 import React, { useCallback } from 'react'
 import styled from 'styled-components/native'
 
-import { useAvailableCredit } from 'features/home/services/useAvailableCredit'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
+import { useAvailableCredit } from 'features/user/helpers/useAvailableCredit'
 import { analytics } from 'libs/firebase/analytics'
 import { useShowReview } from 'libs/hooks/useShowReview'
 import { formatToFrenchDecimal } from 'libs/parsers'
