+++ src/features/favorites/components/Favorite.native.test.tsx
@@ -6,7 +6,7 @@ import { navigate } from '__mocks__/@react-navigation/native'
 import { api } from 'api/api'
 import { ExpenseDomain, FavoriteResponse, SubcategoryIdEnum, UserProfileResponse } from 'api/gen'
 import { initialFavoritesState } from 'features/favorites/context/reducer'
-import { Credit } from 'features/home/services/useAvailableCredit'
+import { Credit } from 'features/user/helpers/useAvailableCredit'
 import { env } from 'libs/environment'
 import { EmptyResponse } from 'libs/fetch'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
