+++ src/features/favorites/components/FavoritesResults.tsx
@@ -17,7 +17,7 @@ import {
   sortByIdDesc,
 } from 'features/favorites/helpers/sorts'
 import { FavoriteSortBy } from 'features/favorites/types'
-import { useAvailableCredit } from 'features/home/services/useAvailableCredit'
+import { useAvailableCredit } from 'features/user/helpers/useAvailableCredit'
 import { useGeolocation, GeoCoordinates } from 'libs/geolocation'
 import { useIsFalseWithDelay } from 'libs/hooks/useIsFalseWithDelay'
 import {
