+++ src/features/identityCheck/pages/Stepper.tsx
@@ -5,7 +5,6 @@ import styled, { useTheme } from 'styled-components/native'
 import { extractApiErrorMessage } from 'api/apiHelpers'
 import { MaintenancePageType, SubscriptionStep } from 'api/gen'
 import { useAuthContext } from 'features/auth/AuthContext'
-import { hasOngoingCredit } from 'features/home/services/useAvailableCredit'
 import { StepButton } from 'features/identityCheck/atoms/StepButton'
 import { FastEduconnectConnectionRequestModal } from 'features/identityCheck/components/FastEduconnectConnectionRequestModal'
 import { QuitIdentityCheckModal } from 'features/identityCheck/components/QuitIdentityCheckModal'
@@ -15,6 +14,7 @@ import { useSetSubscriptionStepAndMethod } from 'features/identityCheck/useSetCu
 import { useSubscriptionSteps } from 'features/identityCheck/useSubscriptionSteps'
 import { getStepState } from 'features/identityCheck/utils/getStepState'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
+import { hasOngoingCredit } from 'features/user/helpers/useAvailableCredit'
 import { analytics } from 'libs/firebase/analytics'
 import { ButtonTertiaryWhite } from 'ui/components/buttons/ButtonTertiaryWhite'
 import { Li } from 'ui/components/Li'
