+++ src/features/identityCheck/pages/confirmation/IdentityCheckHonor.tsx
@@ -5,13 +5,13 @@ import styled, { useTheme } from 'styled-components/native'
 
 import { extractApiErrorMessage } from 'api/apiHelpers'
 import { useAuthContext } from 'features/auth/AuthContext'
-import { hasOngoingCredit } from 'features/home/services/useAvailableCredit'
 import { usePostHonorStatement } from 'features/identityCheck/api/api'
 import { CenteredTitle } from 'features/identityCheck/atoms/CenteredTitle'
 import { Declaration } from 'features/identityCheck/atoms/Declaration'
 import { PageWithHeader } from 'features/identityCheck/components/layout/PageWithHeader'
 import { useSubscriptionNavigation } from 'features/identityCheck/useSubscriptionNavigation'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
+import { hasOngoingCredit } from 'features/user/helpers/useAvailableCredit'
 import { amplitude } from 'libs/amplitude'
 import { QueryKeys } from 'libs/queryKeys'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
