+++ src/features/search/pages/modals/PriceModal/PriceModal.tsx
@@ -7,7 +7,6 @@ import { useTheme } from 'styled-components'
 import { v4 as uuidv4 } from 'uuid'
 
 import { useAuthContext } from 'features/auth/AuthContext'
-import { useAvailableCredit } from 'features/home/services/useAvailableCredit'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { FilterSwitchWithLabel } from 'features/search/components/FilterSwitchWithLabel/FilterSwitchWithLabel'
@@ -19,6 +18,7 @@ import { makeSearchPriceSchema } from 'features/search/helpers/schema/makeSearch
 import { SectionTitle } from 'features/search/helpers/titles'
 import { useLogFilterOnce } from 'features/search/helpers/useLogFilterOnce'
 import { SearchState, SearchView } from 'features/search/types'
+import { useAvailableCredit } from 'features/user/helpers/useAvailableCredit'
 import { formatToFrenchDecimal } from 'libs/parsers'
 import { Banner } from 'ui/components/Banner'
 import { Form } from 'ui/components/Form'
